requirements.txt
apache-airflow == 2.10.2
google-api-python-client == 2.138.0
pandas == 2.1.4
youtube-transcript-api == 0.6.2
yt-dlp == 2024.8.6
